feat(auth): add withAuth and withPublic HOCs for route protection
Implement higher-order components `withAuth` and `withPublic` to handle authentication and public route protection. Move login logic to a dedicated `/login` page and wrap it with `withPublic`. Wrap the home page with `withAuth` to ensure only authenticated users can access it.
